버그 수정.

+++ utils/utils.py
@@ -5,6 +5,7 @@ import torch
 import psutil
 import torchvision.transforms as T
 
+from sys import platform
 from numpy import dtype
 from types import SimpleNamespace
 
@@ -53,18 +54,26 @@ class ParameterServer():
         with self.lock:
             self.weight = weigth
             
-def kill_processes():
-    WORKER_PORTS = [ p.worker_port ]
-    LEARNER_PORTS = [ p.learner_port, p.learner_port+1  ]
-    
-    for port in WORKER_PORTS+LEARNER_PORTS:
-        os.system(f'taskkill /f /pid {port}')
-        
-    parent = psutil.Process( os.getppid() )
-    for child in parent.children(recursive=True):  # or parent.children() for recursive=False
-        child.kill()
-    parent.kill()
-    
-    
-if __name__ == "__main__":
-    kill_processes()
+def kill_processes():    
+    # python main.py로 실행된 프로세스를 찾음 
+    for proc in psutil.process_iter(): 
+        try: # 프로세스 이름, PID값 가져오기 
+            processName = proc.name() 
+            processID   = proc.pid 
+            if processName[:6] == "python": # 윈도우는 python.exe로 올라옴 
+                commandLine = proc.cmdline() 
+                
+                # 동일한 프로세스 확인. code 확인 
+                if 'main.py' in commandLine: 
+                    parent_pid = processID # PID 
+                    parent = psutil.Process(parent_pid) # PID 찾기 
+                    
+                    for child in parent.children(recursive=True): #자식-부모 종료 
+                        child.kill() 
+                        
+                    parent.kill() 
+            else: 
+                print(processName, ' ', proc.cmdline(), ' - ', processID) 
+            
+        except (psutil.NoSuchProcess, psutil.AccessDenied, psutil.ZombieProcess): #예외처리 
+            pass
